Introduce a base class for plugin details
