+++ intellij-plugin-verifier/verifier-repository/src/main/java/com/jetbrains/pluginverifier/plugin/PluginDetailsProviderImpl.kt
@@ -4,92 +4,13 @@
 
 package com.jetbrains.pluginverifier.plugin
 
-import com.jetbrains.plugin.structure.base.plugin.PluginCreationFail
-import com.jetbrains.plugin.structure.base.plugin.PluginCreationSuccess
-import com.jetbrains.plugin.structure.base.problems.PluginProblem
-import com.jetbrains.plugin.structure.base.utils.closeLogged
-import com.jetbrains.plugin.structure.base.utils.closeOnException
-import com.jetbrains.plugin.structure.base.utils.rethrowIfInterrupted
 import com.jetbrains.plugin.structure.intellij.classes.locator.CompileServerExtensionKey
 import com.jetbrains.plugin.structure.intellij.classes.plugin.IdePluginClassesFinder
-import com.jetbrains.plugin.structure.intellij.classes.plugin.IdePluginClassesLocations
 import com.jetbrains.plugin.structure.intellij.plugin.IdePlugin
-import com.jetbrains.plugin.structure.intellij.plugin.IdePluginManager
-import com.jetbrains.plugin.structure.intellij.plugin.StructurallyValidated
-import com.jetbrains.plugin.structure.intellij.problems.UnableToReadPluginFile
 import com.jetbrains.pluginverifier.repository.PluginInfo
-import com.jetbrains.pluginverifier.repository.files.FileLock
 import java.nio.file.Path
 
-/**
- * Main implementation of the [PluginDetailsProvider] that
- * uses the [extractDirectory] for extracting `.zip`-ped plugins.
- */
-class PluginDetailsProviderImpl(private val extractDirectory: Path) : PluginDetailsProvider {
-  private val idePluginManager = IdePluginManager.createManager(extractDirectory)
-
-  override fun providePluginDetails(pluginInfo: PluginInfo, pluginFileLock: FileLock) =
-    pluginFileLock.closeOnException {
-      with(idePluginManager.createPlugin(pluginFileLock.file)) {
-        when (this) {
-          is PluginCreationSuccess -> {
-            readPluginClasses(
-              pluginInfo,
-              plugin,
-              plugin.problems,
-              pluginFileLock
-            )
-          }
-
-          is PluginCreationFail -> {
-            pluginFileLock.closeLogged<FileLock?>()
-            PluginDetailsProvider.Result.InvalidPlugin(pluginInfo, errorsAndWarnings)
-          }
-        }
-      }
-    }
-
-  override fun providePluginDetails(
-    pluginInfo: PluginInfo,
-    idePlugin: IdePlugin
-  ): PluginDetailsProvider.Result {
-    return readPluginClasses(pluginInfo, idePlugin, idePlugin.problems, null)
-  }
-
-  private fun readPluginClasses(
-    pluginInfo: PluginInfo,
-    idePlugin: IdePlugin,
-    warnings: List<PluginProblem>,
-    pluginFileLock: FileLock?
-  ): PluginDetailsProvider.Result {
-    return try {
-      readPluginClasses(pluginInfo, idePlugin)
-        .let { pluginClassesLocations ->
-          PluginDetailsProvider.Result.Provided(
-            PluginDetails(
-              pluginInfo,
-              idePlugin,
-              warnings,
-              pluginClassesLocations,
-              pluginFileLock
-            )
-          )
-        }
-    } catch (e: Exception) {
-      e.rethrowIfInterrupted()
-      val message = e.message ?: e.javaClass.simpleName
-      PluginDetailsProvider.Result.InvalidPlugin(pluginInfo, listOf(UnableToReadPluginFile(message)))
-    }
-  }
-
-  private fun readPluginClasses(
-    pluginInfo: PluginInfo,
-    idePlugin: IdePlugin
-  ): IdePluginClassesLocations =
-    IdePluginClassesFinder
-      .findPluginClasses(idePlugin, additionalKeys = listOf(CompileServerExtensionKey))
-
-  private val IdePlugin.problems: List<PluginProblem>
-    get() = if (this is StructurallyValidated) this.problems else emptyList()
-
+class PluginDetailsProviderImpl(extractDirectory: Path) : AbstractPluginDetailsProvider(extractDirectory) {
+  override fun readPluginClasses(pluginInfo: PluginInfo, idePlugin: IdePlugin) =
+    IdePluginClassesFinder.findPluginClasses(idePlugin, additionalKeys = listOf(CompileServerExtensionKey))
 }
